FEAT: Add function to check if a monotoncurve exists
When two polygonal curves P, Q and an epsilon value is given, the function check if there is a monotone curve that penetrates the free space defined by the polygonal curves and the epsilon value. It uses the method of DFS and traveises the points of L and B which are the member variables of the class FreeSpace in pre-determined order.