UPDATE: Modify checkIfMonotoneCurveExists and remove recursiveCheck
Modified the function checkIfMonotoneCurveExists and removed the function recursiveCheck. Original implementation using recursive method was not effective. Implemented the method that was presented in the paper to check if there is a monotone curve which penetrates the free space. Added new member variables L_R, B_R, which represents reachable area resides in free space.